Refactor IAuditable
