+++ Apps/Common/src/Common.csproj
@@ -44,6 +44,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Audit\" />
-    <Folder Include="Audit\Models\" />
+    <Folder Include="Database\" />
+    <Folder Include="Database\Models\" />
   </ItemGroup>
 </Project>
