+++ Apps/Common/src/Database/Models/Audit.cs
@@ -13,25 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.DIN.Models
+namespace HealthGateway.Common.Database.Models
 {
     using System;
-    using System.ComponentModel.DataAnnotations;
 
-    public abstract class AuditableEntity : IAuditable
+    public class Audit
     {
-        [Required]
-        [MaxLength(30)]
-        public string CreatedBy { get; set; }
-
-        [Required]
-        public DateTime CreatedDateTime { get; set; }
-
-        [Required]
-        [MaxLength(30)]
-        public string UpdatedBy { get; set; }
-
-        [Required]
-        public DateTime UpdatedDateTime { get; set; }
+        public Audit()
+        {
+        }
     }
 }
